# Fixed reindexing arith with duplicates