@TomAugspurger 2020-07-16
Having trouble with ASV right now. Things seem OK in this small benchmark though

```python
In [15]: df1 = pd.DataFrame(np.random.randn(1000, 500))

In [16]: df2 = pd.DataFrame(np.random.randn(1000, 501))

In [17]: %timeit df1 + df2
# master
3.8 ms ± 115 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# PR
3.77 ms ± 104 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```